Revert "Assert that attempt subsegments are truly children of execution subseg"
This reverts commit 10f33bde42c5c09215bc353fa7ad1201e28219ef.

While these assertions are super-necessary, it turns out they can never
work because the X-Ray `Segment` struct doesn't export any of the fields
we really need to check. (Assertions should be running against
`Segment.parent`, which is the actual direct parent subsegment, not
`Segment.ParentSegment`, which is the root segment.