Prevent disaggregating attempt sub-segments in X-Ray console
When you tag both the execution subsegment and its child attempt
subsegments with namespace "remote", it results in the X-Ray console
showing a bifurcated timeline graph where the parent of the execution
subsegment is also the parent of the attempt subsegment.

By removing the "remote" namespace from the attempts, we ensure the
attempts stay nested below the execution in all views.